[GVN] Try to be more careful about handling constant zero, NFCI
Very similar to the previous work, such as "Fix a crash bug w/non-integral
pointers and memtransfers". But try to be more precise about what should
and should not be permitted, and consolidate all users into calling
canCoerceMustAliasedTypeToLoad (since that is the predicate that
VNCoercion will check).

Casting a zero should always be permitted, though the recent work to add
scalable vectors does not appear to yet support that (as the code is
assuming it was always safely check that the size for whether the
alignment layout are compatible, but scalable vector now may raise
exceptions there).

Differential Revision: https://reviews.llvm.org/D59661